Add special type for available fields
