+++ site/gatsby-site/server/fields/incidents.ts
@@ -1,6 +1,6 @@
 import { GraphQLFieldConfigMap, GraphQLFloat, GraphQLInputObjectType, GraphQLInt, GraphQLList, GraphQLNonNull, GraphQLObjectType, GraphQLString } from "graphql";
 import { allow } from "graphql-shield";
-import { generateMutationFields, generateQueryFields, getListRelationshipExtension, getListRelationshipResolver, incidentEmbedding } from "../utils";
+import { generateMutationFields, generateQueryFields, getListRelationshipExtension, getListRelationshipResolver } from "../utils";
 import { ReportType } from "./reports";
 import { getMongoDbQueryResolver } from "graphql-to-mongodb";
 import { Context } from "../interfaces";
@@ -9,6 +9,25 @@ import { Incident, Report } from "../generated/graphql";
 import { isAdmin } from "../rules";
 import { EntityType } from "./entities";
 import { UserType } from "./users";
+
+
+export const incidentEmbedding = (reports: Report[]) => {
+    reports = reports.filter((report) => report.embedding);
+    return reports.length == 0
+        ? null
+        : {
+            vector: reports
+                .map((report) => report.embedding!.vector)
+                .reduce(
+                    (sum, vector) => vector!.map((component, i) => component + sum[i]),
+                    Array(reports[0].embedding!.vector!.length).fill(0)
+                )
+                .map((component) => component / reports.length),
+
+            from_reports: reports.map((report) => report.report_number),
+        };
+};
+
 
 const EmbeddingType = new GraphQLObjectType({
     name: 'IncidentEmbedding',
